Merged from https://code.google.com/p/vldocking-plus/source/detail?r=6
Bug Fix: When closing a dockable from expand panel, i.e. if the close button is clicked when in auto-hide mode, the autohide button is not removed and button zone is not set to invisible.

Bug Fix: Remove autohide buttons when a dockable is set to UNHIDE/DOCKED.
Caching autohide buttons doesn't play nice if the dockables are not cached. For example, if using spring-rcp and page caching is disabled then close operation on a dockable results in the page component being closed and unregistered from the docking content. Since the autohide buttons keeps a reference to the dock key it is associated with, it needs to follow the lifecycle of the dockable, i.e. remove autohide button if the dockable is removed.

Minor changes: Don't set lower minimum sizes. If the user can set the dockable size to an invisble component (really small size), how would he/she resize it again ? It would make sense to use whatever minimum size the component has. The docking framework should not dictate the min/max size of a component.